+++ CHANGELOG.rst
@@ -1,5 +1,6 @@
 1.4.82 (unreleased)
 ------------------------
+* Bug fix: check if validate_email raises error, in case email server can't be contacted
 
 1.4.81 (2015-06-08)
 ------------------------